test actor critic
